# Use builtin_inline_p to skip a frame of C methods